+++ CHANGELOG.md
@@ -1,3 +1,30 @@
+Version 0.2.2
+-------------
+
+**Improvements**
+
+- Bugfix: Fix nullpointer on setting process stream flags.
+
+Version 0.2.1
+-------------
+
+**Improvements**
+
+- Bugfix: Reading `jarsigner verify` output streams got stuck on some Android App bundles.
+
+Version 0.2.0
+-------------
+
+**Improvements**
+
+- Feature: Add new command `android-app-bundle`
+- Feature: Include [Bundletool](https://developer.android.com/studio/command-line/bundletool) jar in the distribution
+- Bugfix: Gracefully handle Xcodeproj exceptions on saving React Native iOS project code signing settings 
+
+**Deprecations**
+
+- Add deprecation notice to `universal-apk` command
+
 Version 0.1.9
 -------------